feat: improve script path handling; refactor: improve code structure
