+++ pyproject.toml
@@ -68,7 +68,7 @@ runtime-evaluated-decorators = ["attrs.define"]
     "S101",  # Use of assert detected
     "SLF001",  # Private member accessed
 ]
-"scripts/_docs_includes.py" = [
+"src/_docs_includes.py" = [
    "RUF001",  # String contains ambiguous Unicode character - `×`
 ]
 "src/types.py" = [
