+++ src/static_data/map_index.py
